requirements.txt
Django==2.2
djangorestframework==3.11.0
django-rest-auth==0.9.5
django-allauth==0.41.0
djangorestframework-jwt==1.11.0
psycopg2==2.8.5
